MSI: Modularity improvement
