+++ app/code/Magento/InventorySales/Plugin/InventoryApi/StockRepository/LoadSalesChannelsOnGetPlugin.php
@@ -5,18 +5,18 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventorySales\Plugin\Inventory\StockRepository;
+namespace Magento\InventorySales\Plugin\InventoryApi\StockRepository;
 
 use Magento\Framework\Exception\StateException;
-use Magento\InventoryApi\Api\Data\StockSearchResultsInterface;
+use Magento\InventoryApi\Api\Data\StockInterface;
 use Magento\InventoryApi\Api\StockRepositoryInterface;
 use Magento\InventorySales\Model\GetAssignedSalesChannelsForStockInterface;
 use Psr\Log\LoggerInterface;
 
 /**
- * Load Sales Channels for Stock on GetList method of StockRepositoryInterface
+ * Load Sales Channels for Stock on Get method of StockRepositoryInterface
  */
-class LoadSalesChannelsOnGetListPlugin
+class LoadSalesChannelsOnGetPlugin
 {
     /**
      * @var GetAssignedSalesChannelsForStockInterface
@@ -41,42 +41,26 @@ class LoadSalesChannelsOnGetListPlugin
     }
 
     /**
-     * Enrich the given Stock Objects with the assigned sales channel entities
+     * Enrich the given Stock Object with the assigned sales channel entities
      *
      * @param StockRepositoryInterface $subject
-     * @param StockSearchResultsInterface $stockSearchResults
-     * @return StockSearchResultsInterface
+     * @param StockInterface $stock
+     * @return StockInterface
      * @throws StateException
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
-    public function afterGetList(
-        StockRepositoryInterface $subject,
-        StockSearchResultsInterface $stockSearchResults
-    ): StockSearchResultsInterface {
-        try {
-            return $this->doAfterGetList($stockSearchResults);
-        } catch (\Exception $e) {
-            $this->logger->error($e->getMessage());
-            throw new StateException(__('Could not load Sales Channels for Stock'), $e);
-        }
-    }
-
-    /**
-     * @param StockSearchResultsInterface $stockSearchResults
-     * @return StockSearchResultsInterface
-     */
-    private function doAfterGetList(StockSearchResultsInterface $stockSearchResults): StockSearchResultsInterface
+    public function afterGet(StockRepositoryInterface $subject, StockInterface $stock): StockInterface
     {
-        $stocks = [];
-        foreach ($stockSearchResults->getItems() as $stock) {
+        try {
             $salesChannels = $this->getAssignedSalesChannelsForStock->execute((int)$stock->getStockId());
 
             $extensionAttributes = $stock->getExtensionAttributes();
             $extensionAttributes->setSalesChannels($salesChannels);
             $stock->setExtensionAttributes($extensionAttributes);
-            $stocks[] = $stock;
+            return $stock;
+        } catch (\Exception $e) {
+            $this->logger->error($e->getMessage());
+            throw new StateException(__('Could not load Sales Channels for Stock'), $e);
         }
-        $stockSearchResults->setItems($stocks);
-        return $stockSearchResults;
     }
 }
